Fragments replaced properly - at last!
Turns out that FragmentTransaction.remove/replace() would not replace a
fragment that has been added via XML. It has to be added
programatically, through a transaction. This is what we do here.

The prefs still don't do anything, but the fragment
transitions/replacements finally do what one would expect them to!

Also now, the GPSUpdatesFragment posts its own Toolbar in the ActionBar.
In a hindsight this may not have been necessary, but is probably OK.